+++ packages/admin/tests/Feature/Http/Livewire/Pages/Collections/CollectionGroupIndexTest.php
@@ -31,6 +31,8 @@ class CollectionGroupIndexTest extends TestCase
     /** @test */
     public function cant_view_page_without_permission()
     {
+        $this->setupRolesPermissions();
+
         $staff = Staff::factory()->create([
             'admin' => false,
         ]);
@@ -44,15 +46,13 @@ class CollectionGroupIndexTest extends TestCase
     /** @test */
     public function can_view_page_with_correct_permission()
     {
+        $this->setupRolesPermissions();
+
         $staff = Staff::factory()->create([
             'admin' => false,
         ]);
 
-        $staff->permissions()->createMany([
-            [
-                'handle' => 'catalogue:manage-collections',
-            ],
-        ]);
+        $staff->givePermissionTo('catalogue:manage-collections');
 
         $this->actingAs($staff, 'staff');
 
@@ -63,13 +63,10 @@ class CollectionGroupIndexTest extends TestCase
     /** @test */
     public function will_redirect_to_collection_group_if_exists()
     {
+        $this->setupRolesPermissions();
         $staff = Staff::factory()->create();
 
-        $staff->permissions()->createMany([
-            [
-                'handle' => 'catalogue:manage-collections',
-            ],
-        ]);
+        $staff->givePermissionTo('catalogue:manage-collections');
 
         $this->actingAs($staff, 'staff');