+++ packages/admin/tests/Feature/Http/Livewire/Pages/Customers/CustomersIndexTest.php
@@ -30,6 +30,8 @@ class CustomersIndexTest extends TestCase
     /** @test */
     public function cant_view_page_without_permission()
     {
+        $this->setupRolesPermissions();
+
         $staff = Staff::factory()->create([
             'admin' => false,
         ]);
@@ -43,15 +45,13 @@ class CustomersIndexTest extends TestCase
     /** @test */
     public function can_view_page_with_correct_permission()
     {
+        $this->setupRolesPermissions();
+
         $staff = Staff::factory()->create([
             'admin' => false,
         ]);
 
-        $staff->permissions()->createMany([
-            [
-                'handle' => 'catalogue:manage-customers',
-            ],
-        ]);
+        $staff->givePermissionTo('catalogue:manage-customers');
 
         $this->actingAs($staff, 'staff');