+++ packages/admin/CHANGELOG.md
@@ -20,6 +20,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Improved `wire:key` on product associations when editing a product to avoid duplicate keys causing DOM diffing issues.
+- Use `spatie/laravel-permission` for hub authorization
+- Removed `Staff->authorize()`
 - Show third-party permissions in staff setting
 - Menu Link/Group/Section now is hidden if no permission given